requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2==2.9.9
jinja2==3.1.4
python-multipart==0.0.9
bcrypt==4.2.0
python-dotenv==1.0.1
